# ARROW-11601: [C++][Python][Dataset] expose Parquet pre-buffer option