@dianaclarke 2021-02-17
@ursabot please benchmark