@lidavidm 2021-02-18
Ah - to be clear, calling PreBuffer doesn't actually block until the data is read! It computes the coalesced read ranges, then asynchronously submits read requests for each of those ranges. The blocking happens later on, when the Parquet reader reads a chunk of the file.