@lidavidm 2021-02-18
But the alternate ScanTask approach makes sense to me. Let me implement that.