@bkietz 2021-02-18
If the blocking situation I described wouldn't happen, then pre buffering at top level is definitely preferable; we don't want to lose row group level parallelism if it's not necessary