@westonpace 2021-02-18
So `PreBuffer` needs to be called once and only once on a reader?  It seems this could be simplified if `PreBuffer` didn't immediately start reading (and waited for the first request) although I suppose that may only push the flag further down.