@lidavidm 2021-02-18
> So `PreBuffer` needs to be called once and only once on a reader? It seems this could be simplified if `PreBuffer` didn't immediately start reading (and waited for the first request) although I suppose that may only push the flag further down.

It can be called multiple times, but all readers created from the file prior to the next call must be discarded first.

Also, you may want PreBuffer to start work immediately: you could pipeline decoding and I/O that way (start prebuffering immediately, then create and sequentially decode the row groups).