@lidavidm 2021-02-18
I agree the API is a bit complex, but it was intentionally rather low-level while we worked out the uses for it.