@westonpace 2021-02-18
Ok, I think I get it now.  Let's pretend we are outside of S3, there is 1 file, 3 row groups, and 3 columns.  Three scan tasks will be generated.  Task 1 needs RG1C1, RG1C2, RG1C3.  Task 2 needs RG2C1, RG2C2, RG2C3.  Task 3 needs RG3C1, RG3C2, RG3C3.

Prebuffer will be called asking for all 9 blocks and it will then issue three reads in parallel (instead of the 9 reads that would otherwise be issued) RG1, RG2, and RG3.

If there are only 3 columns in the file, is it possible Prebuffer would coalesce this all into one read?  In that case wouldn't all three tasks be blocked until the entire file is read, preventing the ability for task 1 to start running as soon as RG1 is issued?